Runge Kutta time step algorithm
A Runge-Kutta time step algorithm implmented in both python and C++,
that can take any derivative in y'=f(y,t) format and calculate the time
steps (ie integrals of the differential equation). Returns the results
as a list of x and y values. The python version uses Pylab to plot the
results